refactor(reviews): extract SubmittedReviewView to its own file
SubmittedReviewView now accepts rubricTemplate/values/rationales/overallComment
as props instead of reading from the ReviewFormContext, so callers outside the
review screen can render a read-only review. FormShell and TotalScoreCard
live alongside it and are re-imported by the editable form.
